modified:   classes/include/num_q.h 	modified:   classes/include/num_z.h 	modified:   classes/include/polynomial.h 	modified:   classes/num_z/abs_bool.cpp 	modified:   classes/num_z/attrib.cpp 	modified:   classes/num_z/bool.cpp 	modified:   classes/num_z/io.cpp 	modified:   classes/num_z/num_z.cpp 	modified:   classes/num_z/op_atrdiv.cpp 	modified:   classes/num_z/op_atrmod.cpp 	modified:   classes/num_z/op_atrmul.cpp 	modified:   classes/num_z/op_atrsub.cpp 	modified:   classes/num_z/op_atrsum.cpp 	modified:   classes/num_z/opdiv.cpp 	modified:   classes/num_z/opmod.cpp 	modified:   classes/num_z/opmul.cpp 	modified:   classes/num_z/oppow.cpp 	modified:   classes/num_z/opsub.cpp 	modified:   classes/num_z/opsum.cpp 	modified:   classes/num_z/private_ops.cpp 	modified:   classes/polynomial/polynomial.cpp

+++ classes/include/polynomial.h
@@ -18,8 +18,8 @@ class polynomial{
 	private:
 		tripla _termo[500];
 		bool _vars[255];
-		unsigned long long _n_termos;
-		unsigned int _n_vars;
+		uint64_t _n_termos;
+		uint32_t _n_vars;
 		
 	public:
 		polynomial();
