+++ classes/num_z/bool.cpp
@@ -11,21 +11,21 @@ bool num_z::operator==(const num_z &a) const {
 bool num_z::operator==(const int &a) const {
 	if(this->_blocks ^ 1) return false;
 	if(this->_sign ^ (a < 0) && this->_num[0] != 0) return false;
-	return (a < 0)?this->_num[0] == (unsigned long long)-a:this->_num[0] == (unsigned long long)a;
+	return (a < 0)?this->_num[0] == (uint64_t)-a:this->_num[0] == (uint64_t)a;
 }
 
-bool num_z::operator==(const unsigned int &a) const {
+bool num_z::operator==(const uint32_t &a) const {
 	if(this->_blocks ^ 1 || this->_sign & 1) return false;
 	return this->_num[0] == a;
 }
 
-bool num_z::operator==(const long long &a) const {
+bool num_z::operator==(const int64_t &a) const {
 	if(this->_blocks ^ 1) return false;
 	if(this->_sign ^ (a < 0) && this->_num[0] != 0) return false;
-	return (a < 0)?this->_num[0] == (unsigned long long)-a:this->_num[0] == (unsigned long long)a;
+	return (a < 0)?this->_num[0] == (uint64_t)-a:this->_num[0] == (uint64_t)a;
 }
 
-bool num_z::operator==(const unsigned long long &a) const {
+bool num_z::operator==(const uint64_t &a) const {
 	if(this->_blocks > 2 || this->_sign & 1) return false;
 	if(this->_blocks == 1) return this->_num[0] == a;
 	if(a > _BLOCK_SIZE_64_) return ((this->_num[1] == 1) && (this->_num[0] == a - _BLOCK_SIZE_64_ - 1)); 
@@ -58,25 +58,25 @@ bool num_z::operator>(const num_z &a) const {
 bool num_z::operator>(const int &a) const {
 	if(this->_sign == 0)
 		if(a < 0 || this->_blocks ^ 1) return true;
-		else return this->_num[0] > (unsigned long long)a;
+		else return this->_num[0] > (uint64_t)a;
 	else if(a >= 0 || this->_blocks ^ 1) return false;
-		else return this->_num[0] < (unsigned long long)-a;
+		else return this->_num[0] < (uint64_t)-a;
 }
 
-bool num_z::operator>(const unsigned int &a) const {
+bool num_z::operator>(const uint32_t &a) const {
 	if(this->_sign & 1) return false;
 	if(this->_blocks ^ 1) return true;
 	return this->_num[0] > a;
 }
-bool num_z::operator>(const long long &a) const{
+bool num_z::operator>(const int64_t &a) const{
 	if(this->_sign == 0)
 		if(a < 0 || this->_blocks ^ 1) return true;
-		else return this->_num[0] > (unsigned long long)a;
+		else return this->_num[0] > (uint64_t)a;
 	else if(a >= 0 || this->_blocks ^ 1) return false;
-		else return this->_num[0] < (unsigned long long)-a;
+		else return this->_num[0] < (uint64_t)-a;
 }
 
-bool num_z::operator>(const unsigned long long &a) const {
+bool num_z::operator>(const uint64_t &a) const {
 	if(this->_sign & 1) return false;
 	if(this->_blocks > 2) return true;
 	if(a > _BLOCK_SIZE_64_)
@@ -88,30 +88,30 @@ bool num_z::operator>(const unsigned long long &a) const {
 
 /*bool num_z::operator<(const num_z &) const;
 bool num_z::operator<(const int &) const;
-bool num_z::operator<(const unsigned int &) const;
-bool num_z::operator<(const long long &) const;
-bool num_z::operator<(const unsigned long long &) const;
+bool num_z::operator<(const uint32_t &) const;
+bool num_z::operator<(const int64_t &) const;
+bool num_z::operator<(const uint64_t &) const;
 bool num_z::operator>=(const num_z &) const;
 bool num_z::operator>=(const int &) const;
-bool num_z::operator>=(const unsigned int &) const;
-bool num_z::operator>=(const long long &) const;
-bool num_z::operator>=(const unsigned long long &) const;
+bool num_z::operator>=(const uint32_t &) const;
+bool num_z::operator>=(const int64_t &) const;
+bool num_z::operator>=(const uint64_t &) const;
 bool num_z::operator<=(const num_z &) const;
 bool num_z::operator<=(const int &) const;
-bool num_z::operator<=(const unsigned int &) const;
-bool num_z::operator<=(const long long &) const;
-bool num_z::operator<=(const unsigned long long &) const;
+bool num_z::operator<=(const uint32_t &) const;
+bool num_z::operator<=(const int64_t &) const;
+bool num_z::operator<=(const uint64_t &) const;
 bool num_z::operator!=(const num_z &) const;
 */
 
 bool num_z::operator!=(const int &a) const {
 	if(this->_blocks > 1) return true;
 	if((a < 0) ^ this->_sign) return true;
-	return (unsigned long long)a ^ this->_num[0];
+	return (uint64_t)a ^ this->_num[0];
 }
 /*	
-bool num_z::operator!=(const unsigned int &) const;
-bool num_z::operator!=(const long long &) const;
-bool num_z::operator!=(const unsigned long long &) const;
+bool num_z::operator!=(const uint32_t &) const;
+bool num_z::operator!=(const int64_t &) const;
+bool num_z::operator!=(const uint64_t &) const;
 
 */
