+++ classes/num_z/op_atrdiv.cpp
@@ -2,8 +2,8 @@
 
 /*
 num_z & num_z::operator/=(const num_z &);
-num_z & num_z::operator/=(const long long &);
-num_z & num_z::operator/=(const unsigned long long &);
+num_z & num_z::operator/=(const int64_t &);
+num_z & num_z::operator/=(const uint64_t &);
 num_z & num_z::operator/=(const int &);
-num_z & num_z::operator/=(const unsigned int &);
+num_z & num_z::operator/=(const uint32_t &);
 */
